+++ packages/ui/ToggleItem/ToggleItem.js
@@ -15,28 +15,34 @@ import kind from '@enact/core/kind';
 import EnactPropTypes from '@enact/core/internal/prop-types';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
-import React from 'react';
+import React, {Fragment} from 'react';
 
 import ComponentOverride from '../ComponentOverride';
+import Slottable from '../Slottable';
 import Toggleable from '../Toggleable';
 import Touchable from '../Touchable';
 
 import componentCss from './ToggleItem.module.less';
 
 // eslint-disable-next-line
-const iconCreator = (position) => ({disabled, icon, iconComponent, selected, iconPosition}) => {
+const iconCreator = (position) => ({disabled, icon, iconComponent, selected, iconPosition, beforeItem}) => {
 	if (iconPosition === position) {
 		return (
-			<ComponentOverride
-				component={iconComponent}
-				disabled={disabled}
-				selected={selected}
-			>
-				{icon}
-			</ComponentOverride>
+			<Fragment>
+				{iconPosition === 'before' && beforeItem ? beforeItem : null}
+				<ComponentOverride
+					component={iconComponent}
+					disabled={disabled}
+					selected={selected}
+				>
+					{icon}
+				</ComponentOverride>
+				{iconPosition === 'after' && beforeItem ? beforeItem : null}
+			</Fragment>
 		);
 	}
 };
+
 
 /**
  * A minimally styled toggle item without any behavior, ripe for extension.
@@ -50,6 +56,7 @@ const ToggleItemBase = kind({
 	name: 'ui:ToggleItem',
 
 	propTypes: /** @lends ui/ToggleItem.ToggleItemBase.prototype */ {
+
 		/**
 		 * The main content of the toggle item.
 		 *
@@ -85,6 +92,35 @@ const ToggleItemBase = kind({
 		 */
 		iconComponent: EnactPropTypes.componentOverride.isRequired,
 
+		/**
+		 * Nodes to be inserted after `children`.
+		 *
+		 * If nothing is specified, nothing, not even an empty container, is rendered in this place.
+		 *
+		 * @type {Node}
+		 * @public
+		 */
+		afterText: PropTypes.node,
+
+		/**
+		 * Nodes to be inserted before `ToggleIcon`.
+		 *
+		 * If nothing is specified, nothing, not even an empty container, is rendered in this place.
+		 *
+		 * @type {Node}
+		 * @public
+		 */
+		beforeItem: PropTypes.node,
+
+		/**
+		 * Nodes to be inserted before `children`.
+		 *
+		 * If nothing is specified, nothing, not even an empty container, is rendered in this place.
+		 *
+		 * @type {Node}
+		 * @public
+		 */
+		beforeText: PropTypes.node,
 		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
 		 * corresponding internal Elements and states of this component.
@@ -128,52 +164,13 @@ const ToggleItemBase = kind({
 		iconPosition: PropTypes.oneOf(['before', 'after']),
 
 		/**
-		 * Called when the toggle item is toggled. Developers should generally use `onToggle` instead.
-		 *
-		 * @type {Function}
-		 * @public
-		 */
-		onTap: PropTypes.func,
-
-		/**
-		 * Called when the toggle item is toggled.
-		 *
-		 * @type {Function}
-		 * @param {Object} event
-		 * @param {String} event.selected - Selected value of item.
-		 * @param {*} event.value - Value passed from `value` prop.
-		 * @public
-		 */
-		onToggle: PropTypes.func,
-
-		/**
 		 * Applies the provided `icon`.
 		 *
 		 * @type {Boolean}
 		 * @default false
 		 * @public
 		 */
-		selected: PropTypes.bool,
-
-		/**
-		 * Nodes to be inserted after `children`.
-		 *
-		 * If nothing is specified, nothing, not even an empty container, is rendered in this place.
-		 *
-		 * @type {Node}
-		 * @public
-		 */
-		slotImageAfter: PropTypes.node,
-
-		/**
-		 * Nodes to be inserted after `children`.
-		 *
-		 * If nothing is specified, nothing, not even an empty container, is rendered in this place.
-		 *
-		 * @type {Node}
-		 * @public
-		 */
-		slotImageBefore: PropTypes.node,
+		onTap: PropTypes.func,
 
 		/**
 		 * The value that will be sent to the `onToggle` handler.
@@ -182,6 +179,8 @@ const ToggleItemBase = kind({
 		 * @default null
 		 * @public
 		 */
+		onToggle: PropTypes.func,
+		selected: PropTypes.bool,
 		value: PropTypes.any
 	},
 
@@ -189,8 +188,6 @@ const ToggleItemBase = kind({
 		disabled: false,
 		iconPosition: 'before',
 		selected: false,
-		slotImageAfter: null,
-		slotImageBefore: null,
 		value: null
 	},
 
@@ -205,7 +202,8 @@ const ToggleItemBase = kind({
 		slotAfter: iconCreator('after')
 	},
 
-	render: ({component: Component, css, children, selected, slotImageAfter, slotImageBefore, ...rest}) => {
+	render: ({component: Component, afterText, beforeText, css, children, selected, ...rest}) => {
+		delete rest.beforeItem;
 		delete rest.iconComponent;
 		delete rest.iconPosition;
 		delete rest.value;
@@ -217,9 +215,9 @@ const ToggleItemBase = kind({
 				css={css}
 				aria-checked={selected}
 			>
-				{slotImageBefore}
+				{beforeText}
 				{children}
-				{slotImageAfter}
+				{afterText}
 			</Component>
 		);
 	}
@@ -236,9 +234,11 @@ const ToggleItemBase = kind({
  * @public
  */
 const ToggleItemDecorator = compose(
+	Slottable({slots: ['beforeText', 'afterText']}),
 	Toggleable({toggleProp: 'onTap'}),
 	Touchable
 );
+
 
 /**
  * An unstyled item with built-in support for toggling.